oz-backend-django/pyproject.toml
[tool.poetry]
name = "oz-backend-django"
version = "0.1.0"
description = ""
authors = ["INSEOP KIM <inseop0813@gmail.com>"]
readme = "README.md"
packages = [{include = "oz_backend_django"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.4"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
